--- Assets/UniGLTF/Editor/UniGLTF/EditorImporterContext.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a1c0c4a153cd28949be45976d229f585
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
